# swagger.yaml
openapi: 3.0.3
info:
  title: Claude Review Demo API
  version: 0.1.0
servers:
  - url: http://localhost:3000/api
paths:
  /auth/login:
    post:
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 6, type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user:
                    type: object
                    properties:
                      id: { type: integer }
                      email: { type: string, format: email }
                      name: { type: string }
                      role: { type: string }
  /users:
    get:
      operationId: listUsers
      parameters:
        - in: query
          name: q
          schema: { type: string, nullable: true }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
        - in: query
          name: sort
          schema: { type: string, enum: [id, name, email], default: id }
        - in: query
          name: order
          schema: { type: string, enum: [asc, desc], default: asc }
      responses:
        '200':
          description: OK
    post:
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, name, password]
              properties:
                email: { type: string, format: email }
                name: { type: string }
                role: { type: string, enum: [admin, user], default: user }
                password: { type: string, minLength: 6 }
      responses:
        '201':
          description: Created
  /users/me:
    get:
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
  /users/{id}:
    delete:
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '204':
          description: No Content
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT